+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-19 17:21:32
-// Commit: style(core): harmonize light and dark theme palette for core
+// Updated at: 2026-06-19 22:56:16
+// Commit: perf(core): optimize memory consumption in core list viewport
 
 class coreComponentState {
-  final String commitRef = '2141';
-  final String lastMessage = 'style(core): harmonize light and dark theme palette for core';
-  final DateTime timestamp = DateTime.parse('2026-06-19T17:21:32+05:00');
+  final String commitRef = '2163';
+  final String lastMessage = 'perf(core): optimize memory consumption in core list viewport';
+  final DateTime timestamp = DateTime.parse('2026-06-19T22:56:16+05:00');
 }